1.完成下载速度统计 2。修改全部删除时通知UI、 3。增加进度双缓冲，但是只能用DrawString显示百分比（没使用） 4。DEMO修改为点下载才构建task对象

+++ Test/MainForm.cs
@@ -22,8 +22,8 @@ namespace CMusicSearch.Test
         MSLRCRunner Finder = new MSLRCRunner(); //搜索对象
         NetworkHelper networkHelp = NetworkHelper.GetNetworkHelperInstance();//网络辅助对象
         DownloadManagement downloadManager = new DownloadManagement(); //下载管理器
-        DownloadMusicTask task = null; //新下载任务
         DataTable downListTable = new DataTable(); //下载列表信息
+        int searchSelectIndex = -1;   //选择的搜索歌曲的序号
         #endregion
 
         public MainForm()
@@ -43,18 +43,33 @@ namespace CMusicSearch.Test
             DataColumn nameCol = new DataColumn("name", typeof(string));
             DataColumn statusCol = new DataColumn("status", typeof(string));
             DataColumn processCol = new DataColumn("process", typeof(int));
+            DataColumn speedCol = new DataColumn("speed", typeof(string));
+            DataColumn sizeCol = new DataColumn("size", typeof(string));
             DataColumn linkCol = new DataColumn("link", typeof(string));
             DataColumn guidCol = new DataColumn("GUID", typeof(Guid));
+            DataColumn taskCol = new DataColumn("task", typeof(DownloadMusicTask)); //存放task
             downListTable.Columns.Add(nameCol);
             downListTable.Columns.Add(statusCol);
             downListTable.Columns.Add(processCol);
+            downListTable.Columns.Add(speedCol);
+            downListTable.Columns.Add(sizeCol);
             downListTable.Columns.Add(linkCol);
             downListTable.Columns.Add(guidCol);
+            downListTable.Columns.Add(taskCol);
             dataGridView3.Columns[0].DataPropertyName = "name";
-            dataGridView3.Columns[1].DataPropertyName = "status";
-            dataGridView3.Columns[2].DataPropertyName = "process";
-            dataGridView3.Columns[3].DataPropertyName = "link";
+            dataGridView3.Columns[1].DataPropertyName = "size";
+            dataGridView3.Columns[2].DataPropertyName = "status";
+            dataGridView3.Columns[3].DataPropertyName = "process";
+            dataGridView3.Columns[4].DataPropertyName = "speed";
+            dataGridView3.Columns[5].DataPropertyName = "link";
+            dataGridView3.Columns[6].DataPropertyName = "GUID";
+            dataGridView3.Columns[7].DataPropertyName = "task";
             dataGridView3.DataSource = downListTable;
+
+            //双缓冲
+            this.DoubleBuffered = true;
+            //this.SetStyle(ControlStyles.DoubleBuffer | ControlStyles.UserPaint | ControlStyles.AllPaintingInWmPaint, true);
+            //this.UpdateStyles();
         }
 
         #region 网络状态
@@ -165,39 +180,53 @@ namespace CMusicSearch.Test
                 float process = 0;
                 if (totalSize != 0)
                     process = (downSize / totalSize) * 100;
-                //if (task.DownloadSpeed != -1)
-                //String.Format("{0}k/s", task.DownloadSpeed / 1024);
 
-                //更新下载信息
+                //更新下载状态
                 DataRow[] updateRow = downListTable.Select("GUID='" + task.DownloadTaskID.ToString() + "'");
+
+                //下载进度和速度,文件大小
                 if (updateRow != null && updateRow.Count() > 0)
                 {
                     updateRow[0]["process"] = (int)process;
+                    string speed = String.Format("{0}KB/S", task.DownloadSpeed / 1024);
+                    if (updateRow[0]["speed"].ToString() != speed)
+                        updateRow[0]["speed"] = speed;
+                    float size = task.FileSize / 1024 / 1024;
+                    string sizeString = size.ToString("F2") + "M";
+                    if (updateRow[0]["size"].ToString() != sizeString)
+                        updateRow[0]["size"] = sizeString;
                 }
 
+                //下载状态
                 if (task.DownloadStatus == DownloadStatus.ST_READY_DOWNLOAD)
                 {
-                    updateRow[0]["status"] = "准备下载";
+                    if (updateRow[0]["status"].ToString() != "准备下载")
+                        updateRow[0]["status"] = "准备下载";
                 }
                 else if (task.DownloadStatus == DownloadStatus.ST_IS_DOWNLOAD)
                 {
-                    updateRow[0]["status"] = "下载中";
+                    if (updateRow[0]["status"].ToString() != "下载中")
+                        updateRow[0]["status"] = "下载中";
                 }
                 else if (task.DownloadStatus == DownloadStatus.ST_WAIT_DOWNLOAD)
                 {
-                    updateRow[0]["status"] = "等待下载";
+                    if (updateRow[0]["status"].ToString() != "等待下载")
+                        updateRow[0]["status"] = "等待下载";
                 }
                 else if (task.DownloadStatus == DownloadStatus.ST_ERROR_DOWNLOAD)
                 {
-                    updateRow[0]["status"] = "下载失败";
+                    if (updateRow[0]["status"].ToString() != "下载失败")
+                        updateRow[0]["status"] = "下载失败";
                 }
                 else if (task.DownloadStatus == DownloadStatus.ST_STOP_DOWNLOAD)
                 {
-                    updateRow[0]["status"] = "暂停下载";
+                    if (updateRow[0]["status"].ToString() != "暂停下载")
+                        updateRow[0]["status"] = "暂停下载";
                 }
                 else if (task.DownloadStatus == DownloadStatus.ST_NONE)
                 {
-                    updateRow[0]["status"] = string.Empty;
+                    if (updateRow[0]["status"].ToString() != string.Empty)
+                        updateRow[0]["status"] = string.Empty;
                 }
             }
             catch (System.Exception ex)
@@ -237,7 +266,7 @@ namespace CMusicSearch.Test
 
                 //对下载的文件进行操作
                 FileManager.FileDownloadOver(result.MusicSavePath, result.MusicConfigPath, result.DownloadStatus);
-
+                updateRow[0]["speed"] = string.Empty;
             }
             catch (System.Exception ex)
             {
@@ -253,15 +282,7 @@ namespace CMusicSearch.Test
         {
             if (e.RowIndex > -1)
             {
-                task = new DownloadMusicTask();
-                task.DownloadUrl = dataGridView1["MusicUrl", e.RowIndex].Value.ToString();
-                task.MusicName = dataGridView1["MusicName", e.RowIndex].Value.ToString();
-                task.MusicSavePath = FileManager.GetSavaPath(task.MusicName, dataGridView1["MusicFormat", e.RowIndex].Value.ToString());
-                if (string.IsNullOrEmpty(task.MusicSavePath))
-                {
-                    task = null;
-                    MessageBox.Show("无法保存文件");
-                }
+                searchSelectIndex = e.RowIndex;
             }
         }
 
@@ -271,23 +292,34 @@ namespace CMusicSearch.Test
         /// </summary>
         private void btnDownload_Click(object sender, EventArgs e)
         {
-            if (task != null)
+            if (searchSelectIndex != -1)
             {
-                //设置下载列表
+                //设置下载任务
+                DownloadMusicTask task = new DownloadMusicTask();
+                task.DownloadUrl = dataGridView1["MusicUrl", searchSelectIndex].Value.ToString();
+                task.MusicName = dataGridView1["MusicName", searchSelectIndex].Value.ToString();
+                task.MusicSavePath = FileManager.GetSavaPath(task.MusicName, dataGridView1["MusicFormat", searchSelectIndex].Value.ToString());
+                if (string.IsNullOrEmpty(task.MusicSavePath))
+                {
+                    task = null;
+                    MessageBox.Show("无法保存文件");
+                }
                 task.DownloadTaskID = Guid.NewGuid();
 
-
+                //添加到下载列表
                 DataRow row = downListTable.NewRow();
-                row[0] = task.MusicName;
-                row[1] = string.Empty;
-                row[2] = 0;
-                row[3] = task.DownloadUrl;
-                row[4] = task.DownloadTaskID;
+                row["name"] = task.MusicName;
+                row["size"] = string.Empty;
+                row["status"] = string.Empty;
+                row["process"] = 0;
+                row["speed"] = string.Empty;
+                row["link"] = task.DownloadUrl;
+                row["GUID"] = task.DownloadTaskID;
+                row["task"] = task;
                 downListTable.Rows.Add(row);
 
-
+                //开始下载
                 downloadManager.RunWorkerAsync(task);
-
             }
         }
 
@@ -298,6 +330,11 @@ namespace CMusicSearch.Test
         int downloadIndex = -1;
         private void dataGridView3_CellMouseClick(object sender, DataGridViewCellMouseEventArgs e)
         {
+            if (e.Button != MouseButtons.Right)
+            {
+                downloadIndex = -1;
+                return;
+            }
             downloadIndex = e.RowIndex;
             pauseAllToolStripMenuItem.Enabled = true;
             deleteAllToolStripMenuItem.Enabled = true;
@@ -334,14 +371,14 @@ namespace CMusicSearch.Test
         /// <param name="e"></param>
         private void pauseToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            Guid taskID = (Guid)downListTable.Rows[downloadIndex]["GUID"];
+            DownloadMusicTask task = (DownloadMusicTask)downListTable.Rows[downloadIndex]["task"];
             if (downListTable.Rows[downloadIndex]["status"].ToString() == "下载中")
             {
-                downloadManager.StopAsync(taskID);
+                downloadManager.StopAsync(task.DownloadTaskID);
             }
             if (downListTable.Rows[downloadIndex]["status"].ToString() == "暂停下载")
             {
-                //downloadManager.RunWorkerAsync(taskID);
+                downloadManager.RunWorkerAsync(task);
             }
         }
 
@@ -380,7 +417,7 @@ namespace CMusicSearch.Test
 
         private void button2_Click(object sender, EventArgs e)
         {
-            System.Diagnostics.Process.Start("explorer.exe", FileManager.DOWNLOAD_DIR); 
+            System.Diagnostics.Process.Start("explorer.exe", FileManager.DOWNLOAD_DIR);
         }
 
     }
